Rework editor convenience; QoL; goal and start

+++ js/Editor.js
@@ -1,3 +1,6 @@
+import { ObjectID } from "./ObjectID.js";
+import { IDToType } from "./ObjectID.js";
+
 console.log("Editor.js loaded");
 
 export class Editor {
@@ -12,11 +15,14 @@ export class Editor {
 
         this.mouseX = 0;
         this.mouseY = 0;
+        this.mouseDown = false;
+        this.mouseButton = null;
 
         this.selectedObjects = [];
         this.isSelecting = false;
 
-        this.showHelp = true;
+        this.history = [];
+        this.redoStack = [];
 
         this.bindInputs();
     }
@@ -25,6 +31,36 @@ export class Editor {
         window.addEventListener("mousemove", e => {
             this.mouseX = e.clientX;
             this.mouseY = e.clientY;
+
+            if (!this.game.editorMode) {
+                return;
+            }
+
+            if (this.mouseDown) {
+                if (this.mouseButton === 0) {
+                    if (e.ctrlKey) {
+                        return;
+                    }
+
+                    const obj = this.getObjectAtCursor();
+
+                    if (!obj) {
+                        this.placeObject();
+                    }
+                }
+
+                if (this.mouseButton === 2) {
+                    if (e.ctrlKey) {
+                        return;
+                    }
+
+                    const obj = this.getObjectAtCursor();
+
+                    if (obj) {
+                        this.deleteObject();
+                    }
+                }
+            }
         });
 
         window.addEventListener("mousedown", e => {
@@ -32,26 +68,46 @@ export class Editor {
                 return;
             }
 
-            if (e.button === 2) {
-                this.deleteObject();
+            this.mouseDown = true;
+            this.mouseButton = e.button;
 
-                return;
-            }
-
+            // Place Objects
             if (e.button === 0) {
-                const clickedObj = this.handleSelection(e);
+                const obj = this.getObjectAtCursor();
+
+                if (obj) {
+                    this.handleSelection(e);
+
+                    return;
+                }
 
                 if (e.ctrlKey) {
                     return;
                 }
 
-                if (clickedObj) {
+                this.placeObject();
+            }
+
+            // Delete Objects
+            if (e.button === 2) {
+                const obj = this.getObjectAtCursor();
+
+                if (e.ctrlKey) {
                     return;
                 }
 
-                this.placeObject();
+                if (obj) {
+                    this.deleteObject();
+                }
+
+                return;
             }
         });
+
+        window.addEventListener("mouseup", e => {
+            this.mouseDown = false;
+            this.mouseButton = null;
+        })
 
         window.addEventListener("keydown", e => {
             if (!this.game.editorMode) {
@@ -66,71 +122,103 @@ export class Editor {
             if (e.code === "Digit2") {
                 this.selectedType = "spike";
             }
+            if (e.code === "Digit3") {
+                this.selectedType = "start";
+            }
+            if (e.code === "Digit4") {
+                this.selectedType = "goal";
+            }
 
             // Movement
 
             const step = e.shiftKey ? 0.1 : 1;
 
-            if (e.code === "KeyW") {
-                if (!e.shiftKey) {
-                    this.camera.y -= this.grid;
+            if (!e.ctrlKey) {
+                if (e.code === "KeyW") {
+                    if (!e.shiftKey) {
+                        this.camera.y -= this.grid;
+                    }
+
+                    this.moveSelected(0, -step);
+                }
+                if (e.code === "KeyS") {
+                    if (!e.shiftKey) {
+                        this.camera.y += this.grid;
+                    }
+
+                    this.moveSelected(0, step);
+                }
+                if (e.code === "KeyA") {
+                    if (!e.shiftKey) {
+                        this.camera.x -= this.grid;
+                    }
+
+                    this.moveSelected(-step, 0);
+                }
+                if (e.code === "KeyD") {
+                    if (!e.shiftKey) {
+                        this.camera.x += this.grid;
+                    }
+
+                    this.moveSelected(step, 0);
                 }
 
-                this.moveSelected(0, -step);
-            }
-            if (e.code === "KeyS") {
-                if (!e.shiftKey) {
-                    this.camera.y += this.grid;
+                // Rotation
+
+                if (e.code === "ArrowUp") {
+                    this.setSelectedRotation(0);
                 }
-
-                this.moveSelected(0, step);
-            }
-            if (e.code === "KeyA") {
-                if (!e.shiftKey) {
-                    this.camera.x -= this.grid;
+                if (e.code === "ArrowRight") {
+                    this.setSelectedRotation(90);
                 }
-
-                this.moveSelected(-step, 0);
-            }
-            if (e.code === "KeyD") {
-                if (!e.shiftKey) {
-                    this.camera.x += this.grid;
+                if (e.code === "ArrowDown") {
+                    this.setSelectedRotation(180);
                 }
-
-                this.moveSelected(step, 0);
-            }
-
-            // Rotation
-
-            if (e.code === "ArrowLeft") {
-                this.rotateSelected(-90);
-            }
-            if (e.code === "ArrowRight") {
-                this.rotateSelected(90);
+                if (e.code === "ArrowLeft") {
+                    this.setSelectedRotation(270);
+                }
             }
 
             // Misc
 
-            if (e.code === "KeyC") {
+            if (e.code === "KeyC" && e.ctrlKey) {
                 this.copySelected();
             }
 
-            if (e.code === "KeyV") {
+            if (e.code === "KeyZ" && e.ctrlKey) {
+                this.undo();
+            }
+
+            if (e.code === "KeyY" && e.ctrlKey) {
+                this.redo();
+            }
+
+            if (e.code === "Backspace") {
                 this.deleteSelected();
             }
 
-            if (e.code === "KeyX") {
+            if (e.code === "KeyS" && e.ctrlKey) {
+                e.preventDefault();
+
                 const data = this.exportLevel();
                 navigator.clipboard.writeText(data);
                 alert("Level data copied to clipboard!");
             }
 
-            if (e.code === "KeyZ") {
+            if (e.code === "KeyL" && e.ctrlKey) {
+                e.preventDefault();
+
                 const json = prompt("Paste level JSON:");
                 
                 if (json) {
                     this.importLevel(json);
                 }
+            }
+
+            if (e.code === "KeyR" && e.ctrlKey && e.shiftKey) {
+                e.preventDefault();
+
+                this.resetEditor();
             }
 
             if (e.code === "Escape") {
@@ -155,12 +243,36 @@ export class Editor {
         });
     }
 
+    getObjectAtCursor() {
+        const worldX = this.camera.screenToWorldX(this.mouseX);
+        const worldY = this.camera.screenToWorldY(this.mouseY);
+    
+        return this.level.find(o => (
+            worldX >= o.x * this.grid &&
+            worldX <  (o.x * this.grid + this.grid) &&
+            worldY >= o.y * this.grid &&
+            worldY <  (o.y * this.grid + this.grid)
+        ));
+    }
+    
     placeObject() {
+        this.saveHistory();
+
         const worldX = this.camera.screenToWorldX(this.mouseX);
         const worldY = this.camera.screenToWorldY(this.mouseY);
 
         const gx = Math.floor(worldX / this.grid);
         const gy = Math.floor(worldY / this.grid);
+
+        if (this.selectedType === "start" && this.level.some(o => o.type === "start")) {
+            return;
+        }
+
+        /*
+        if (this.selectedType === "goal" && this.level.some(o => o.type === "start")) {
+            return;
+        }
+        */
 
         const newObj = {
             type: this.selectedType,
@@ -174,28 +286,61 @@ export class Editor {
         this.selectedObjects = [newObj];
 
         this.game.loadLevel(this.level);
+        this.saveToLocalStorage();
     }
 
     deleteObject() {
+        this.saveHistory();
+
+        const obj = this.getObjectAtCursor();
+        
+        if (!obj) {
+            return;
+        }
+
+        this.level = this.level.filter(o => o !== obj);
+        this.selectedObjects = [];
+        this.game.loadLevel(this.level);
+        this.saveToLocalStorage();
+
+        /*
         const worldX = this.camera.screenToWorldX(this.mouseX);
         const worldY = this.camera.screenToWorldY(this.mouseY);
 
-        const gx = Math.floor(worldX / this.grid);
-        const gy = Math.floor(worldY / this.grid);
+        //const gx = Math.floor(worldX / this.grid);
+        //const gy = Math.floor(worldY / this.grid);
 
-        this.level = this.level.filter(obj => !(obj.x === gx && obj.y === gy));
+        this.level = this.level.filter(obj => {
+            const left = obj.x * this.grid;
+            const right = left + this.grid;
+            const top = obj.y * this.grid;
+            const bottom = top + this.grid;
+
+            return !(worldX >= left && worldX < right && worldY >= top && worldY < bottom);
+        });
+
+        this.selectedObjects = [];
+
+        //this.level = this.level.filter(obj => !(obj.x === gx && obj.y === gy));
 
         this.game.loadLevel(this.level);
+        this.saveToLocalStorage();
+        */
     }
 
     handleSelection(e) {
         const worldX = this.camera.screenToWorldX(this.mouseX);
         const worldY = this.camera.screenToWorldY(this.mouseY);
 
-        const gx = Math.floor(worldX / this.grid);
-        const gy = Math.floor(worldY / this.grid);
+        //const gx = Math.floor(worldX / this.grid);
+        //const gy = Math.floor(worldY / this.grid);
 
-        const obj = this.level.find(o => o.x === gx && o.y === gy);
+        const obj = this.level.find(o =>
+            worldX >= o.x * this.grid &&
+            worldX <  (o.x * this.grid + this.grid) &&
+            worldY >= o.y * this.grid &&
+            worldY <  (o.y * this.grid + this.grid)
+        );
 
         if (!obj) {
             this.selectedObjects = [];
@@ -219,6 +364,8 @@ export class Editor {
     }
 
     moveSelected(dx, dy) {
+        this.saveHistory();
+
         for (const obj of this.selectedObjects) {
             obj.x += dx;
             obj.y += dy;
@@ -228,17 +375,24 @@ export class Editor {
         }
 
         this.game.loadLevel(this.level);
+        this.saveToLocalStorage();
     }
 
-    rotateSelected(angle) {
+    setSelectedRotation(angle) {
+        this.saveHistory();
+    
         for (const obj of this.selectedObjects) {
-            obj.rotation = (obj.rotation || 0) + angle;
+            obj.rotation = angle;
         }
-
+    
         this.game.loadLevel(this.level);
+        this.saveToLocalStorage();
     }
+    
 
     copySelected() {
+        this.saveHistory();
+
         if (this.selectedObjects.length === 0) {
             return;
         }
@@ -260,6 +414,7 @@ export class Editor {
         this.selectedObjects = clones;
 
         this.game.loadLevel(this.level);
+        this.saveToLocalStorage();
     }
 
     deleteSelected() {
@@ -272,6 +427,7 @@ export class Editor {
         this.selectedObjects = [];
 
         this.game.loadLevel(this.level);
+        this.saveToLocalStorage();
     }
 
     draw(ctx) {
@@ -279,7 +435,6 @@ export class Editor {
         this.drawPreview(ctx);
         this.drawSelectedType(ctx);
         this.drawSelection(ctx);
-        this.drawHelp(ctx);
     }
 
     drawGrid(ctx) {
@@ -347,12 +502,12 @@ export class Editor {
         ctx.fillText("Selected: ", 10, 25);
 
         if (this.selectedType === "block") {
-            ctx.fillStyle = "rgba(0,150,255,0.8)";
+            ctx.fillStyle = "rgba(0,150,255,0.5)";
             ctx.fillRect(10, 35, this.grid * 0.6, this.grid * 0.6);
         }
 
         if (this.selectedType === "spike") {
-            ctx.fillStyle = "rgba(255,50,50,0.8)";
+            ctx.fillStyle = "rgba(255,50,50,0.5)";
 
             const size = this.grid * 0.6;
             const x = 10;
@@ -366,9 +521,20 @@ export class Editor {
             ctx.fill();
         }
 
+        if (this.selectedType === "start") {
+            ctx.fillStyle = "rgba(255,0,0,0.5)";
+            ctx.fillRect(10, 35, this.grid * 0.6, this.grid * 0.6);
+        }
+
+        if (this.selectedType === "goal") {
+            ctx.fillStyle = "rgba(0,255,0,0.5)";
+            ctx.fillRect(10, 35, this.grid * 0.6, this.grid * 0.6);
+        }
+
         ctx.restore();
     }
 
+    /*
     drawHelp(ctx) {
         if (!this.showHelp) {
             return;
@@ -400,13 +566,102 @@ export class Editor {
     
         ctx.restore();
     }
+    */
+
+    saveHistory() {
+        const snapshot = JSON.parse(JSON.stringify(this.level));
+        this.history.push(snapshot);
+
+        this.redoStack = [];
+    }
+
+    undo() {
+        if (this.history.length === 0) {
+            return;
+        }
+
+        const current = JSON.parse(JSON.stringify(this.level));
+        this.redoStack.push(current);
+
+        this.level = this.history.pop();
+
+        this.selectedObjects = [];
+
+        this.game.loadLevel(this.level);
+        this.saveToLocalStorage();
+    }
+
+    redo() {
+        if (this.redoStack.length === 0) {
+            return;
+        }
+
+        const current = JSON.parse(JSON.stringify(this.level));
+        this.history.push(current);
+
+        this.level = this.redoStack.pop();
+
+        this.selectedObjects = [];
+
+        this.game.loadLevel(this.level);
+        this.saveToLocalStorage();
+    }
 
     exportLevel() {
-        return JSON.stringify(this.level);
+        const objects = this.level.map(obj => {
+            const id = ObjectID[obj.type];
+            const arr = [id, obj.x, obj.y];
+    
+            if (obj.rotation !== undefined) {
+                arr.push(obj.rotation);
+            }
+    
+            return arr;
+        });
+    
+        const levelData = {
+            bg: this.game.backgroundColor,
+            mode: this.game.mode,
+            objects
+        };
+    
+        return JSON.stringify(levelData);
     }
 
     importLevel(json) {
-        this.level = JSON.parse(json);
+        const data = JSON.parse(json);
+    
+        this.game.backgroundColor = data.bg ?? "#000";
+        this.game.mode = data.mode ?? 1;
+    
+        const parsed = data.objects.map(arr => {
+            const [id, x, y, rotation] = arr;
+    
+            return {
+                type: IDToType[id],
+                x,
+                y,
+                rotation: rotation ?? 0
+            };
+        });
+    
+        this.level = parsed;
+    
+        this.game.loadLevel(parsed);
+        this.saveToLocalStorage();
+    }
+
+    saveToLocalStorage() {
+        localStorage.setItem("level", this.exportLevel());
+    }
+
+    resetEditor() {
+        this.level = [];
+        this.selectedObjects = [];
+        this.history = [];
+        this.redoStack = [];
+
         this.game.loadLevel(this.level);
+        this.saveToLocalStorage();
     }
 }
